move harness to wanthttp
