+++ src/wantjob/wanthttp/harness.go
@@ -1,4 +1,4 @@
-package wantqemu
+package wanthttp
 
 import (
 	"context"
@@ -8,7 +8,6 @@ import (
 	"go.brendoncarroll.net/state/cadata"
 	"wantbuild.io/want/src/internal/streammux"
 	"wantbuild.io/want/src/wantjob"
-	"wantbuild.io/want/src/wantjob/wanthttp"
 )
 
 // MainRW connects to a wanthttp API served on rw and calls fn with a job context for the server.
@@ -17,11 +16,11 @@ func MainRW(rw io.ReadWriter, fn func(jc wantjob.Ctx, s cadata.Getter, input []b
 	hc := &http.Client{
 		Transport: streammux.NewRoundTripper(mux),
 	}
-	wc := wanthttp.NewClient(hc, "")
+	wc := NewClient(hc, "")
 	jc := wantjob.Ctx{
 		Context: context.Background(),
 		System:  wc,
-		Dst:     wc.Store(wanthttp.CurrentStore),
+		Dst:     wc.Store(CurrentStore),
 	}
 	input, inputStore, err := wc.GetInput(context.Background())
 	if err != nil {
